refactor: split golden seed
