+++ packages/database/src/seed-golden/constants.ts
@@ -1,0 +1,7 @@
+import type { GoldenTenants } from './types';
+
+export const TENANTS: GoldenTenants = {
+  MK: 'tenant_mk',
+  KS: 'tenant_ks',
+  PILOT: 'pilot-mk',
+};
