+++ packages/database/test/seed-golden.test.ts
@@ -3,11 +3,18 @@ import test from 'node:test';
 
 test('buildSeededMembershipCardIdentifiers returns deterministic values for seeded subscriptions', async () => {
   const seedGoldenModule = (await import('../src/seed-golden')) as {
+    seedGolden?: (config: unknown) => Promise<void>;
     buildSeededMembershipCardIdentifiers?: (
       subscriptionId: string,
       tenantId: string
     ) => { cardNumber: string; qrCodeToken: string };
   };
+
+  assert.equal(
+    typeof seedGoldenModule.seedGolden,
+    'function',
+    'expected seed-golden to export the public seeder entrypoint'
+  );
 
   assert.equal(
     typeof seedGoldenModule.buildSeededMembershipCardIdentifiers,
@@ -36,4 +43,48 @@ test('buildSeededMembershipCardIdentifiers returns deterministic values for seed
     buildSeededMembershipCardIdentifiers('sub_ks_a_1', 'tenant_ks'),
     'expected helper output to stay stable across repeated calls'
   );
+});
+
+test('seedGolden orchestrator preserves sequential module-call order', async () => {
+  const { SEED_GOLDEN_STEP_ORDER, runSeedGoldenSteps } = await import('../src/seed-golden/index');
+  const expectedOrder = [
+    'cleanup',
+    'tenants',
+    'branches',
+    'users',
+    'agent-assignments',
+    'memberships',
+    'agent-settings',
+    'claims',
+    'leads',
+    'tracking-tokens',
+    'member-counters',
+  ];
+  const calls: string[] = [];
+  const makeStep = (name: string) => async () => {
+    calls.push(name);
+  };
+  const originalLog = console.log;
+
+  console.log = () => {};
+  try {
+    await runSeedGoldenSteps({} as never, {
+      cleanup: makeStep('cleanup'),
+      tenants: makeStep('tenants'),
+      branches: makeStep('branches'),
+      users: makeStep('users'),
+      agentAssignments: makeStep('agent-assignments'),
+      memberships: makeStep('memberships'),
+      agentSettings: makeStep('agent-settings'),
+      claims: makeStep('claims'),
+      leads: makeStep('leads'),
+      trackingTokens: makeStep('tracking-tokens'),
+      memberCounters: makeStep('member-counters'),
+    });
+  } finally {
+    console.log = originalLog;
+  }
+
+  assert.deepEqual(SEED_GOLDEN_STEP_ORDER, expectedOrder);
+  assert.deepEqual(calls, expectedOrder);
 });
